Deprecate Apstra's aos_* modules
These modules don't work with AOS 2.1 or higher.
They will be replaced non-upstreamed modules
