+++ lib/ansible/modules/network/aos/_aos_blueprint_virtnet.py
@@ -19,7 +19,7 @@
 #
 
 ANSIBLE_METADATA = {'metadata_version': '1.1',
-                    'status': ['preview'],
+                    'status': ['deprecated'],
                     'supported_by': 'community'}
 
 
@@ -29,6 +29,10 @@ module: aos_blueprint_virtnet
 author: Damien Garros (@dgarros)
 version_added: "2.3"
 short_description: Manage AOS blueprint parameter values
+deprecated:
+    removed_in: "2.9"
+    why: This module does not support AOS 2.1 or later
+    alternative: See new modules at U(https://www.ansible.com/ansible-apstra).
 description:
  - Apstra AOS Blueprint Virtual Network module let you manage your Virtual Network easily.
    You can create access, define and delete Virtual Network by name or by using a JSON / Yaml file.
