+++ lib/ansible/modules/network/aos/_aos_device.py
@@ -19,7 +19,7 @@
 #
 
 ANSIBLE_METADATA = {'metadata_version': '1.1',
-                    'status': ['preview'],
+                    'status': ['deprecated'],
                     'supported_by': 'community'}
 
 
@@ -29,6 +29,10 @@ module: aos_device
 author: Damien Garros (@dgarros)
 version_added: "2.3"
 short_description: Manage Devices on AOS Server
+deprecated:
+    removed_in: "2.9"
+    why: This module does not support AOS 2.1 or later
+    alternative: See new modules at U(https://www.ansible.com/ansible-apstra).
 description:
   - Apstra AOS Device module let you manage your devices in AOS easily. You can
     approve devices and define in which state the device should be. Currently
